Add sector option to the radar
Radar sector command sets sector size what radar scans. The sectors default is 180 degrees which means full sphere. 90 degrees means half sphere in front of tthe ship and 30 deg is smaller cone in front. The sector/code is always direccted to the ship heading.

The the range will increas to proportional to the size of sector. Halfing the sector size doubles the range.